+++ crates/nautobot-openapi/src/apis/cloud_api.rs
@@ -10,7 +10,7 @@
 
 use reqwest;
 
-use super::{configuration, Error};
+use super::{Error, configuration};
 use crate::apis::ResponseContent;
 
 /// struct for typed errors of method [`cloud_cloud_accounts_bulk_destroy`]