+++ crates/nautobot-openapi/src/apis/extras_api.rs
@@ -10,7 +10,7 @@
 
 use reqwest;
 
-use super::{configuration, Error};
+use super::{Error, configuration};
 use crate::apis::ResponseContent;
 
 /// struct for typed errors of method [`extras_computed_fields_bulk_destroy`]